Split of keywords_importer into kkewords_importer and keywords_extractor
