+++ rfhub2/version.py
@@ -1,1 +1,1 @@
-version = "0.19"
+version = "0.20"
