+++ tests/__init__.py
@@ -4,9 +4,8 @@ from tests.unit.api.endpoints.keywords_tests import KeywordsApiTest
 from tests.unit.api.endpoints.statistics_tests import StatisticsApiTest
 from tests.unit.api.endpoints.version_tests import VersionApiTest
 from tests.unit.cli.api_client import ApiClientTests
-from tests.unit.cli.keywords_importer import KeywordsImporterTests
-from tests.unit.cli.statistics_extractor import StatisticsExtractorTests
-from tests.unit.cli.statistics_importer import StatisticsImporterTests
+from tests.unit.cli.keywords.keywords_importer import KeywordsImporterTests
+from tests.unit.cli.statistics.statistics_extractor import StatisticsExtractorTests
 from tests.unit.db.collection_repo_tests import CollectionRepositoryTest
 from tests.unit.db.collection_tests import CollectionTest
 from tests.unit.db.keyword_repo_tests import KeywordRepositoryTest
